Module libsignal_bridge::ffi

Modules

convert

error

storage

Structs

CallbackError
Represents an error returned by a callback, following the C conventions that 0 means "success".

FfiIdentityKeyStoreStruct

FfiPreKeyStoreStruct

FfiSenderKeyStoreStruct

FfiSessionStoreStruct

FfiSignedPreKeyStoreStruct

Enums

FfiDirection

SignalFfiError
The top-level error type (opaquely) returned to C clients when something goes wrong.

Traits

ArgTypeInfo
Converts arguments from their FFI form to their Rust form.

ResultTypeInfo
Converts result values from their Rust form to their FFI form.

SimpleArgTypeInfo
A simpler interface for ArgTypeInfo for when no local storage is needed.

SizedArgTypeInfo
Converts "sized" arguments from their FFI form to their Rust form.

Functions

box_object⚠

expect_ready
Polls a future once; panics if it is not Ready.

native_handle_cast⚠

native_handle_cast_mut⚠

run_ffi_safe

write_bytearray_to⚠

write_result_to⚠

Type Definitions

SignalFfiResult