Trait libsignal_bridge::ffi::convert::ResultTypeInfo

pub trait ResultTypeInfo: Sized {
    type ResultType;
    fn convert_into(self) -> SignalFfiResult<Self::ResultType>;
}

Converts result values from their Rust form to their FFI form.

ResultTypeInfo is used to implement the bridge_fn macro, but can also be used outside it.

let ffi_result = rust_result.convert_into()?;

Implementers should also see the ffi_result_type macro in convert.rs.

Associated Types

type ResultType

The FFI form of the result (e.g. libc::c_uchar).

Loading content...
Required methods

fn convert_into(self) -> SignalFfiResult<Self::ResultType>

Converts the data in self to the FFI type, similar to try_into().

Loading content...