Trait libsignal_bridge::ffi::convert::SimpleArgTypeInfo

pub trait SimpleArgTypeInfo: Sized {
    type ArgType: Copy;
    fn convert_from(foreign: Self::ArgType) -> SignalFfiResult<Self>;
}

A simpler interface for ArgTypeInfo for when no local storage is needed.

This trait is easier to use when writing FFI functions manually:

impl SimpleArgTypeInfo for Foo {
    type ArgType = isize;
    fn convert_from(foreign: isize) -> SignalFfiResult<Self> {
        // ...
    }
}

let rust_arg = Foo::convert_from(ffi_arg)?;

However, some types do need the full flexibility of ArgTypeInfo.

Associated Types

type ArgType: Copy

The FFI form of the argument (e.g. libc::c_uchar).

Must be Copy to help the compiler optimize out local storage.

Loading content...
Required methods

fn convert_from(foreign: Self::ArgType) -> SignalFfiResult<Self>

Converts the data in foreign to the Rust type.

Loading content...