Trait libsignal_bridge::ffi::SizedArgTypeInfo

pub trait SizedArgTypeInfo: Sized {
    type ArgType;
    fn convert_from(
        foreign: Self::ArgType, 
        size: usize
    ) -> SignalFfiResult<Self>;
}

Converts "sized" arguments from their FFI form to their Rust form.

This is used for buffers and such passed as a base+length pair. Implementing types are usually slices; the ArgType will usually be a pointer.

SizedArgTypeInfo is used to implement the bridge_fn macro for slice-typed arguments, but can also be used outside it.

let rust_arg = Foo::convert_from(ffi_arg, ffi_arg_len)?;

Associated Types

type ArgType

The FFI form of the "base" argument (e.g. *const u8).

Note that the "length" argument type is not customizable; it is always usize (size_t in C).

Loading content...
Required methods

fn convert_from(foreign: Self::ArgType, size: usize) -> SignalFfiResult<Self>

Converts the data in foreign to the Rust type.

Loading content...