Trait libsignal_bridge::jni::convert::SimpleArgTypeInfo

pub trait SimpleArgTypeInfo<'a>: Sized {
    type ArgType: Copy + 'a;
    fn convert_from(
        env: &JNIEnv<'_>, 
        foreign: Self::ArgType
    ) -> SignalJniResult<Self>;
}

A simpler interface for ArgTypeInfo for when no local storage is needed.

This trait is easier to use when writing JNI functions manually:

impl<'a> SimpleArgTypeInfo<'a> for Foo {
    type ArgType = JObject<'a>;
    fn convert_from(env: &JNIEnv, foreign: JObject<'a>) -> SignalJniResult<Self> {
        // ...
    }
}

let rust_arg = Foo::convert_from(env, jni_arg)?;

However, some types do need the full flexibility of ArgTypeInfo.

Associated Types

type ArgType: Copy + 'a

The JNI form of the argument (e.g. jint).

Must be Copy to help the compiler optimize out local storage.

Loading content...
Required methods

fn convert_from(
env: &JNIEnv<'_>,
foreign: Self::ArgType
) -> SignalJniResult<Self>

Converts the data in foreign to the Rust type.

Loading content...