Trait libsignal_bridge::jni::ArgTypeInfo

pub trait ArgTypeInfo<'storage, 'context: 'storage>: Sized {
    type ArgType;
    type StoredType: 'storage;
    fn borrow(
        env: &'context JNIEnv<'_>, 
        foreign: Self::ArgType
    ) -> SignalJniResult<Self::StoredType>;

    fn load_from(
        env: &JNIEnv<'_>, 
        stored: &'storage mut Self::StoredType
    ) -> SignalJniResult<Self>;
}

Converts arguments from their JNI form to their Rust form.

ArgTypeInfo has two required methods: borrow and load_from. The use site looks like this:

let mut jni_arg_borrowed = Foo::borrow(env, jni_arg)?;
let rust_arg = Foo::load_from(env, &mut jni_arg_borrowed)?;

The 'context lifetime allows for borrowed values to depend on the current JNI stack frame; that is, they can be assured that referenced objects will not be GC'd out from under them.

ArgTypeInfo is used to implement the bridge_fn macro, but can also be used outside it.

If the Rust type can be directly loaded from ArgType with no local storage needed, implement SimpleArgTypeInfo instead.

Implementers should also see the jni_arg_type macro in convert.rs.

Associated Types

type ArgType

The JNI form of the argument (e.g. jni::jint).

type StoredType: 'storage

Local storage for the argument (ideally borrowed rather than copied).

Loading content...
Required methods

fn borrow(
env: &'context JNIEnv<'_>,
foreign: Self::ArgType
) -> SignalJniResult<Self::StoredType>

"Borrows" the data in foreign, usually to establish a local lifetime or owning type.

fn load_from(
env: &JNIEnv<'_>,
stored: &'storage mut Self::StoredType
) -> SignalJniResult<Self>

Loads the Rust value from the data that's been stored by borrow().

Loading content...