Trait libsignal_bridge::jni::ResultTypeInfo

pub trait ResultTypeInfo: Sized {
    type ResultType;
    fn convert_into(self, env: &JNIEnv<'_>) -> SignalJniResult<Self::ResultType>;
}

Converts result values from their Rust form to their FFI form.

ResultTypeInfo is used to implement the bridge_fn macro, but can also be used outside it.

let jni_result = rust_result.convert_into(env)?;

Implementers should also see the jni_result_type macro in convert.rs.

Associated Types

type ResultType

The JNI form of the result (e.g. jint).

Loading content...
Required methods

fn convert_into(self, env: &JNIEnv<'_>) -> SignalJniResult<Self::ResultType>

Converts the data in self to the JNI type, similar to try_into().

Loading content...