Trait libsignal_bridge::node::convert::ArgTypeInfo

pub trait ArgTypeInfo<'storage, 'context: 'storage>: Sized {
    type ArgType: Value;
    type StoredType: 'storage;
    fn borrow(
        cx: &mut FunctionContext<'context>, 
        foreign: Handle<'context, Self::ArgType>
    ) -> NeonResult<Self::StoredType>;

    fn load_from(stored: &'storage mut Self::StoredType) -> Self;
}

Converts arguments from their JavaScript form to their Rust form.

ArgTypeInfo has two required methods: borrow and load_from. The use site looks like this:

let mut js_arg_borrowed = Foo::borrow(cx, js_arg)?;
let rust_arg = Foo::load_from(&mut js_arg_borrowed);

The 'context lifetime allows for borrowed values to depend on the current JS stack frame; that is, they can be assured that referenced objects will not be GC'd out from under them.

ArgTypeInfo is used to implement the bridge_fn macro, but can also be used outside it.

If the Rust type can be directly loaded from ArgType with no local storage needed, implement SimpleArgTypeInfo instead.

Associated Types

type ArgType: Value

The JavaScript form of the argument (e.g. JsNumber).

type StoredType: 'storage

Local storage for the argument (ideally borrowed rather than copied).

Loading content...
Required methods

fn borrow(
cx: &mut FunctionContext<'context>,
foreign: Handle<'context, Self::ArgType>
) -> NeonResult<Self::StoredType>

"Borrows" the data in foreign, usually to establish a local lifetime or owning type.

fn load_from(stored: &'storage mut Self::StoredType) -> Self

Loads the Rust value from the data that's been stored by borrow().

Loading content...