Trait libsignal_bridge::node::convert::ResultTypeInfo

pub trait ResultTypeInfo<'a>: Sized {
    type ResultType: Value;
    fn convert_into(
        self, 
        cx: &mut impl Context<'a>
    ) -> JsResult<'a, Self::ResultType>;
}

Converts result values from their Rust form to their JavaScript form.

ResultTypeInfo is used to implement the bridge_fn macro, but can also be used outside it.

let js_result = rust_result.convert_into(&mut cx)?;

Implementers should also see the jni_result_type macro in convert.rs.

Associated Types

type ResultType: Value

The JavaScript form of the result (e.g. JsNumber).

Loading content...
Required methods

fn convert_into(
self,
cx: &mut impl Context<'a>
) -> JsResult<'a, Self::ResultType>

Converts the data in self to the JavaScript type, similar to try_into().

Loading content...