Trait libsignal_bridge::node::convert::SimpleArgTypeInfo

pub trait SimpleArgTypeInfo: Sized + 'static {
    type ArgType: Value;
    fn convert_from(
        cx: &mut FunctionContext<'_>, 
        foreign: Handle<'_, Self::ArgType>
    ) -> NeonResult<Self>;
}

A simpler interface for ArgTypeInfo and AsyncArgTypeInfo for when no separate local storage is needed.

This trait is easier to use when writing Neon functions manually:

impl SimpleArgTypeInfo for Foo {
    type ArgType = JsObject;
    fn convert_from(cx: &mut FunctionContext, _: Handle<JsObject>) -> NeonResult<Self> {
        // ...
    }
}

let rust_arg = Foo::convert_from(&mut cx, js_arg)?;

However, some types do need the full flexibility of ArgTypeInfo or AsyncArgTypeInfo.

Associated Types

type ArgType: Value

The JavaScript form of the argument (e.g. JsNumber).

Loading content...
Required methods

fn convert_from(
cx: &mut FunctionContext<'_>,
foreign: Handle<'_, Self::ArgType>
) -> NeonResult<Self>

Converts the data in foreign to the Rust type.

Loading content...